requirements.txt
attrs==22.1.0
autopep8==1.7.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
exceptiongroup==1.0.0
idna==3.4
iniconfig==1.1.1
lxml==4.9.1
medium-stats==2.2.0
packaging==21.3
pluggy==1.0.0
pycodestyle==2.9.1
pyparsing==3.0.9
pytest==7.2.0
python-dotenv==0.21.0
requests==2.28.1
tomli==2.0.1
urllib3==1.26.12
